// src/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  HexColorField: { input: any; output: any; }
  /**
   * Allows use of a JSON String for input / output from the GraphQL schema.
   *
   * Use of this type is *not recommended* as you lose the benefits of having a defined, static
   * schema (one of the key benefits of GraphQL).
   */
  JSONString: { input: any; output: any; }
  L10nField: { input: any; output: any; }
  MediaUrlField: { input: any; output: any; }
};
// ...
export type IconsVehicle = {
  __typename?: 'IconsVehicle';
  contour?: Maybe<Scalars['MediaUrlField']['output']>;
  contourAlive?: Maybe<Scalars['MediaUrlField']['output']>;
  contourDead?: Maybe<Scalars['MediaUrlField']['output']>;
  default?: Maybe<Scalars['MediaUrlField']['output']>;
  large?: Maybe<Scalars['MediaUrlField']['output']>;
  localSmall?: Maybe<Scalars['String']['output']>;
  medium?: Maybe<Scalars['MediaUrlField']['output']>;
  small?: Maybe<Scalars['MediaUrlField']['output']>;
};

export type IconsVehicleClass = {
  __typename?: 'IconsVehicleClass';
  default?: Maybe<Scalars['MediaUrlField']['output']>;
  elite?: Maybe<Scalars['MediaUrlField']['output']>;
  normal?: Maybe<Scalars['MediaUrlField']['output']>;
  premium?: Maybe<Scalars['MediaUrlField']['output']>;
  special?: Maybe<Scalars['MediaUrlField']['output']>;
};
// ...
export type ItemIcons = {
  __typename?: 'ItemIcons';
  default?: Maybe<Scalars['MediaUrlField']['output']>;
  extra?: Maybe<Scalars['MediaUrlField']['output']>;
};

export type ItemInterface = {
  items?: Maybe<Array<Maybe<Item>>>;
};


export type ItemInterfaceItemsArgs = {
  itemGroup?: InputMaybe<Scalars['String']['input']>;
  itemId?: InputMaybe<Scalars['String']['input']>;
  itemIds?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
  itemType?: InputMaybe<Scalars['String']['input']>;
  lang?: InputMaybe<Scalars['String']['input']>;
};

export type ItemRestrictions = {
  __typename?: 'ItemRestrictions';
  levels?: Maybe<Array<Maybe<Scalars['Int']['output']>>>;
  subtype?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
};

export type ItemType = {
  __typename?: 'ItemType';
  icons?: Maybe<ItemTypeIcons>;
  name?: Maybe<Scalars['ID']['output']>;
  title?: Maybe<Scalars['L10nField']['output']>;
};
// ...
export type Nation = {
  __typename?: 'Nation';
  color?: Maybe<Scalars['HexColorField']['output']>;
  icons?: Maybe<NationIcons>;
  name?: Maybe<Scalars['ID']['output']>;
  title?: Maybe<Scalars['L10nField']['output']>;
};

export type NationIcons = {
  __typename?: 'NationIcons';
  default?: Maybe<Scalars['MediaUrlField']['output']>;
  large?: Maybe<Scalars['MediaUrlField']['output']>;
  medium?: Maybe<Scalars['MediaUrlField']['output']>;
  small?: Maybe<Scalars['MediaUrlField']['output']>;
  tiny?: Maybe<Scalars['MediaUrlField']['output']>;
};
// ...
export type Prices = {
  __typename?: 'Prices';
  credit?: Maybe<Scalars['Int']['output']>;
  gold?: Maybe<Scalars['Int']['output']>;
  xp?: Maybe<Scalars['Int']['output']>;
};
// ...
export type Ttc = {
  __typename?: 'TTC';
  name?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['String']['output']>;
  value?: Maybe<Scalars['String']['output']>;
};
// ...
export type Vehicle = {
  __typename?: 'Vehicle';
  consumables?: Maybe<Array<Maybe<VehicleConsumable>>>;
  description?: Maybe<Scalars['L10nField']['output']>;
  hasGoldPrice?: Maybe<Scalars['Boolean']['output']>;
  icons?: Maybe<IconsVehicle>;
  id?: Maybe<Scalars['ID']['output']>;
  isClan?: Maybe<Scalars['Boolean']['output']>;
  isPremium?: Maybe<Scalars['Boolean']['output']>;
  isSpecial?: Maybe<Scalars['Boolean']['output']>;
  level?: Maybe<Scalars['Int']['output']>;
  modernizations?: Maybe<Array<Maybe<VehicleModernization>>>;
  name?: Maybe<Scalars['String']['output']>;
  nation?: Maybe<Nation>;
  nationName?: Maybe<Scalars['String']['output']>;
  tags?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  title?: Maybe<Scalars['L10nField']['output']>;
  titleShort?: Maybe<Scalars['L10nField']['output']>;
  ttc?: Maybe<Array<Maybe<VehicleTtc>>>;
  type?: Maybe<VehicleType>;
  typeName?: Maybe<Scalars['String']['output']>;
};

export type VehicleConsumable = {
  __typename?: 'VehicleConsumable';
  consumableSlot?: Maybe<Scalars['String']['output']>;
  description?: Maybe<Scalars['L10nField']['output']>;
  group?: Maybe<ItemType>;
  groupName?: Maybe<Scalars['String']['output']>;
  icons?: Maybe<ItemIcons>;
  id?: Maybe<Scalars['ID']['output']>;
  name?: Maybe<Scalars['String']['output']>;
  prices?: Maybe<Prices>;
  restrictions?: Maybe<ItemRestrictions>;
  slot?: Maybe<Scalars['Int']['output']>;
  tags?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  title?: Maybe<Scalars['L10nField']['output']>;
  titleShort?: Maybe<Scalars['L10nField']['output']>;
  ttc?: Maybe<Array<Maybe<Ttc>>>;
  type?: Maybe<ItemType>;
  typeName?: Maybe<Scalars['String']['output']>;
};
// ...
export type VehicleModernization = {
  __typename?: 'VehicleModernization';
  description?: Maybe<Scalars['L10nField']['output']>;
  group?: Maybe<ItemType>;
  groupName?: Maybe<Scalars['String']['output']>;
  icons?: Maybe<ItemIcons>;
  id?: Maybe<Scalars['ID']['output']>;
  modernizationSlot?: Maybe<Scalars['String']['output']>;
  name?: Maybe<Scalars['String']['output']>;
  prices?: Maybe<Prices>;
  restrictions?: Maybe<ItemRestrictions>;
  slot?: Maybe<Scalars['Int']['output']>;
  tags?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  title?: Maybe<Scalars['L10nField']['output']>;
  titleShort?: Maybe<Scalars['L10nField']['output']>;
  ttc?: Maybe<Array<Maybe<Ttc>>>;
  type?: Maybe<ItemType>;
  typeName?: Maybe<Scalars['String']['output']>;
};

export type VehicleTtc = {
  __typename?: 'VehicleTTC';
  description?: Maybe<Scalars['L10nField']['output']>;
  name?: Maybe<Scalars['String']['output']>;
  title?: Maybe<Scalars['L10nField']['output']>;
  unit?: Maybe<Scalars['L10nField']['output']>;
  value?: Maybe<Scalars['Float']['output']>;
};

export type VehicleType = {
  __typename?: 'VehicleType';
  icons?: Maybe<IconsVehicleClass>;
  name?: Maybe<Scalars['ID']['output']>;
  title?: Maybe<Scalars['L10nField']['output']>;
  titleShort?: Maybe<Scalars['L10nField']['output']>;
};